#1 add getAntibodies function and clean up AntobodiesTable component
