+++ applications/areg-portal/frontend/src/components/AntibodiesTable.tsx
@@ -1,27 +1,75 @@
 import React from "react";
+//MUI
 import { DataGrid } from "@mui/x-data-grid";
-import { rows, columns } from "../utils/constants";
-import Container from "@mui/material/Container";
+import {
+  GridRowsProp,
+  GridColDef,
+  GridValueGetterParams,
+} from "@mui/x-data-grid";
 import Box from "@mui/material/Box";
-import CssBaseline from "@mui/material/CssBaseline";
+
+//project imports
+import { getAntibodies } from "../utils/helpers";
+import Content from "./UI/Content";
+
+//use renderCell to render diferent types here:
+// renderCell: (params) => (
+//   <div>
+//     <Typography>{params.value.name}</Typography>
+//     <Typography color="textSecondary">{params.value.title}</Typography>
+//   </div>
+
+function joinNameAndId(params: GridValueGetterParams) {
+  let nameIdJoint = `${params.row.ab_name || ""} ${params.row.ab_id || ""}`;
+  return nameIdJoint;
+}
+
+const rows: GridRowsProp = getAntibodies();
+
+const columns: GridColDef[] = [
+  { field: "ab_name", headerName: "Name", minWidth: 100, flex: 1 },
+  { field: "ab_id", headerName: "ID", minWidth: 100, flex: 1 },
+  {
+    field: "ab_name_id",
+    headerName: "Name & ID",
+    width: 250,
+    valueGetter: joinNameAndId,
+  },
+  { field: "ab_target", headerName: "Target antigen", minWidth: 100, flex: 1 },
+  {
+    field: "target_species",
+    headerName: "Target species",
+    minWidth: 100,
+    flex: 1,
+  },
+  {
+    field: "proper_citation",
+    headerName: "Proper citation",
+    minWidth: 100,
+    flex: 1,
+  },
+  { field: "clonality", headerName: "Clonality", minWidth: 100, flex: 1 },
+  { field: "comments", headerName: "Comments", minWidth: 100, flex: 1 },
+  { field: "clone_id", headerName: "Clone ID", minWidth: 100, flex: 1 },
+  { field: "host", headerName: "Host", minWidth: 100, flex: 1 },
+  { field: "vendor", headerName: "Vendor", minWidth: 100, flex: 1 },
+  { field: "catalog_num", headerName: "Cat Num", minWidth: 100, flex: 1 },
+];
 
 function AntibodiesTable() {
   return (
-    <>
-      <CssBaseline />
-      <Container maxWidth="xl">
-        <Box sx={{ height: "100vh" }}>
-          <DataGrid
-            rows={rows}
-            columns={columns}
-            pageSize={20}
-            rowsPerPageOptions={[20]}
-            checkboxSelection
-            disableSelectionOnClick
-          />
-        </Box>
-      </Container>
-    </>
+    <Content>
+      <Box sx={{ height: "100vh" }}>
+        <DataGrid
+          rows={rows}
+          columns={columns}
+          pageSize={20}
+          rowsPerPageOptions={[20]}
+          checkboxSelection
+          disableSelectionOnClick
+        />
+      </Box>
+    </Content>
   );
 }
