+++ applications/areg-portal/frontend/src/utils/helpers.ts
@@ -1,0 +1,3 @@
+import dataJson from "./data.json";
+
+export const getAntibodies = () => dataJson;
